separated edit function
